* Introduce the PLASSERT and PLASSERT_MSG macros in plerror.h
* Global search-and-replace of assert() to PLASSERT()

* Please use PLASSERT from now on.

+++ plearn/vmat/MultiTaskSeparationSplitter.cc
@@ -83,7 +83,7 @@ void MultiTaskSeparationSplitter::declareOptions(OptionList& ol)
 void MultiTaskSeparationSplitter::build_()
 {
     if (seed != 0)
-        random_gen->manual_seed(seed);
+        random_gen->manual_seed(long(seed));
 
     if(tasks.length() == 0 && dataset)
     {
